requirements.txt
# --- Original BCG chatbot deps ---
requests
pandas
yfinance
anthropic
plotly
streamlit
python-dotenv

# --- RAG layer (LangChain ecosystem) ---
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.0

# Vector store + embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0

# 10-K ingestion
sec-edgar-downloader>=5.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Sparse retrieval
rank-bm25>=0.2.2
